# auth_swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Team SpitFire Events App API
  description: API documentation for All Endpoints
basePath: #/api/auth
schemes:
  - https
paths:
...
  /api/events/{event_id}/comments:
    post:
      summary: Add a comment to an event discussion
      description: Add a comment to an event discussion or fetch all comments for an event.
      # parameters:
      #   - name: event_id
      #     in: path
      #     description: The ID of the event causing the discussion.
      #     required: true
      #     type: string
      #   - name: body
      #     in: body
      #     description: Content of Comment.
      #     required: true
      #     schema:
      #       type: object
      #       properties:
      #         body:
      #           type: string
      #       required:
      #         - body
      parameters:
        - name: event_id
          in: path
          description: The ID of the event causing the discussion.
          required: true
          type: string
        - name: Comment
          in: body
          description: add user id and comment body
          required: true
          schema:
            type: object
            properties:
              user_id:
                type: string
              body:
                type: string
      responses:
        201:
          description: Comment added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      body:
                        type: string
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  error:
                    type: string
    get:
      summary: Get comments for an event
      description: Fetch all comments associated with an event.
      parameters:
        - name: event_id
          in: path
          description: The ID of the event to retrieve comments for.
          required: true
          type: string
      responses:
        200:
          description: Comments fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        body:
                          type: string
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  error:
                    type: string
...
  /api/events:
    post:
      summary: Create a new event
      description: Create a new event.
      parameters:
        - name: Event
          in: body
          description: Event Details
          required: true
          schema:
            type: object
            properties:
              title:
                type: string
              description:
                type: string
              location:
                type: string
              start_date:
                type: string
                format: date
              start_time:
                type: string
                format: time
              end_date:
                type: string
                format: date
              end_time:
                type: string
                format: time
              thumbnail:
                type: string
              creator_id:
                type: string
            required:
              - title
              - description
              - location
              - start_date
              - start_time
              - end_date
              - end_time
              - thumbnail
              - creator_id
      responses:
        201:
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  event:
                    type: object
    get:
      summary: Get all events
      description: Retrieve all events in DB.
      # parameters:
      #   - name: id
      #     in: path
      #     description: The ID of the event to retrieve.
      #     required: true
      #     type: string
      responses:
        200:
          description: Event retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      title:
                        type: string
                      description:
                        type: string
                      location:
                        type: string
                      start_date:
                        type: string
                      start_time:
                        type: string
                      end_date:
                        type: string
                      end_time:
                        type: string
                      thumbnail:
                        type: string
                      creator_id:
                        type: string
        404:
          description: Event Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/events/{id}:
    delete:
      summary: Delete an event
      description: Delete an event by ID.
      parameters:
        - name: id
          in: path
          description: The ID of the event to delete.
          required: true
          type: string
      responses:
        204:
          description: Event deleted successfully

    get:
      summary: Get an event by ID
      description: Get an event by its ID.
      parameters:
        - name: id
          in: path
          description: The ID of the event to retrieve.
          required: true
          type: string
      responses:
        200:
          description: Event retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  # Define the properties you want to include in the response
        404:
          description: Event not found

    put:
      summary: Update event details by ID
      description: Update details of an event by its ID.
      parameters:
        - name: id
          in: path
          description: The ID of the event to update.
          required: true
          type: string
        - name: Event
          in: body
          description: Event Details (Only pass in what you want to update)
          required: true
          schema:
            type: object
            properties:
              title:
                type: string
              description:
                type: string
              location:
                type: string
              start_date:
                type: string
                format: date
              start_time:
                type: string
                format: time
              end_date:
                type: string
                format: date
              end_time:
                type: string
                format: time
              thumbnail:
                type: string
              creator_id:
                type: string
            # required:
            #   - title
            #   - description
            #   - location
            #   - start_date
            #   - start_time
            #   - end_date
            #   - end_time
            #   - thumbnail
            #   - creator
      responses:
        201:
          description: Event updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  event:
                    type: object
...